backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.5.0
bcrypt==4.0.1
Pillow==10.0.1
python-dotenv==1.0.0
werkzeug==2.3.7
marshmallow==3.20.1
pytest==7.4.3
gunicorn==21.2.0
